+++ src/cartridge/vrc6.rs
@@ -697,7 +697,7 @@ mod tests {
     }
 
     fn create_vrc6_mapper(
-        mapper_number: u8,
+        mapper_number: u16,
         prg_rom: Vec<u8>,
         chr_rom: Vec<u8>,
         mirroring: MirroringMode,
